refactor: extract data from logic
